requirements.txt
gradio==3.16.2
timm==0.6.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors
torch==1.12.1
torchvision==0.13.1
numpy
opencv-python==4.6.0.66
opencv-contrib-python==4.5.5.64
opencv-python-headless==4.5.5.64
albumentations==1.3.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
einops==0.3.0
transformers==4.30.2
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.20.0
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==0.6.0
xformers
triton
matplotlib
wandb
pillow
pyiqa
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e .
